Port Player + Bandit entities to use new base entity system to reduce repeated code and clutter.
OOP FTW
