+++ entities.py
@@ -13,7 +13,7 @@ stero = SoundSystem()
 import uuid
 import time
 import random
-
+from utilities.base_entities import Entity, Human
 
 class Tumbleweed(pygame.sprite.Sprite):
     def __init__(self, position=None, direction="left"):
@@ -197,13 +197,11 @@ class Bullet(pygame.sprite.Sprite):
         pass
 
 
-class Player(pygame.sprite.Sprite):
+class Player(Human):
     def __init__(self):
         super(Player, self).__init__()
         self.name = "player"
-        self.id = uuid.uuid4()
 
-        self.idle = True
         self.gun_draw = False
 
         self.gunshot_sound = pygame.mixer.Sound(os.path.join(ASSETS_DIRECTORY, SOUNDS_DIRECTORY, "gunshot.wav"))
@@ -211,25 +209,9 @@ class Player(pygame.sprite.Sprite):
         self.Animation_Idle = Animation(CLINT_SPRITESHEET, [0, 4])
         self.Animation_Walk = Animation(CLINT_SPRITESHEET, [4, 10])
 
-        self.direction = "right"
-
         self.surf = self.Animation_Idle.get_frame(position=0, direction=self.direction)
         self.surf.set_colorkey((255, 255, 255))
         self.rect = self.surf.get_rect()
-
-        self.v = [0, 0]
-        self.acceleration = 10
-        self.gravity = 10
-        self.max_v = [50, 200]
-
-        self.jumping = False
-        self.on_tile = False
-        self.jump_velocity = [-30, 0]
-        self.jump_start = self.rect.center
-        self.max_jump_height = 24
-
-        self.animation_count = 0
-        self.spawned_entities = []
 
     def update(self, GS, keys_pressed):
         self.idle = True
@@ -267,56 +249,6 @@ class Player(pygame.sprite.Sprite):
 
         return GS
 
-    def update_movement(self, GS): # TODO update all of this to just use gamestate
-        dt = GS.dt
-        if self.jumping and (self.jump_start[1] - self.rect.center[1]) >= self.max_jump_height:
-            self.jumping = False
-            self.v[1] = 0
-
-        if self.v[0] > self.max_v[0]:
-            self.v[0] = self.max_v[0]
-        elif self.v[0] < -self.max_v[0]:
-            self.v[0] = -self.max_v[0]
-        if self.v[1] > self.max_v[1]:
-            self.v[1] = self.max_v[1]
-        elif self.v[1] < -self.max_v[1]:
-            self.v[1] = -self.max_v[1]
-
-        if self.v[0] < 0:
-            x = -math.ceil((self.v[0] * -1) * dt)
-        else:
-            x = math.ceil(self.v[0] * dt)
-        y = math.ceil(self.v[1] * dt)
-
-        # cx, cy = [self.rect.center[0] // 16, self.rect.center[1] // 16]  # get the tile the rect is currently in
-        # print(f"The player is currently in block ({cx}, {cy})")
-        # print(f"The x velocity is {self.v[0]}\nThe applied x velocity is {x}")
-
-        self.on_tile = False
-        if x != 0:
-            self.rect.move_ip(x, 0)
-            collisions = GS.GameMap.get_collisions(self.rect, GS.Camera.offset)
-            for collide in collisions:
-                if x < 0:  # colliding with the right of a tile
-                    self.rect.left = collide.rect.right
-                else:  # colliding with the left of a tile
-                    self.rect.right = collide.rect.left
-        if y != 0:
-            self.rect.move_ip(0, y)
-            collisions = GS.GameMap.get_collisions(self.rect, GS.Camera.offset)
-            for collide in collisions:
-                if y > 0:  # standing on top of a tile
-                    self.v[1] = 0
-                    self.rect.bottom = collide.rect.top
-                    self.on_tile = True
-                else:  # hitting bottom of tile
-                    self.rect.top = collide.rect.bottom
-
-    def update_direction(self, direction="right"):
-        if self.direction != direction:
-            self.direction = direction
-            self.surf = pygame.transform.flip(self.surf, True, False)  # horizontal flip: true, vertical: false
-
     def update_animation(self):  # TODO swap frame increment and display so that the first frame is displayed
         if not self.gun_draw:
             if self.idle:
@@ -332,12 +264,6 @@ class Player(pygame.sprite.Sprite):
                 self.gun_draw = False
             self.surf = self.Animation_GunDraw.get_frame(direction=self.direction)
         self.surf.set_colorkey((255, 255, 255))
-
-    def trigger_jump(self):
-        if self.on_tile:
-            self.jumping = True
-            self.v[1] = 0
-            self.jump_start = self.rect.center
 
     def fire_gun(self):
         spos = self.rect.center
@@ -356,14 +282,10 @@ class Player(pygame.sprite.Sprite):
     def trigger_gunfire(self):
         self.gun_draw = True
 
-    def on_hit(self):
-        print("PLAYER HIT")
 
-
-class Bandit(pygame.sprite.Sprite):
+class Bandit(Human):
     def __init__(self, start_pos, goal=None, hostile=True):
         super(Bandit, self).__init__()
-        self.id = uuid.uuid4()
         self.name = "bandit"
         self.goal = goal
         self.hostile = hostile
@@ -372,21 +294,16 @@ class Bandit(pygame.sprite.Sprite):
         self.Animation_Walk = Animation("bandit-spritesheet.png", [4, 10])
         self.Animation_GunDraw = Animation("bandit_gun_draw.png", [0, 6])
 
-        self.idle = True
         self.gun_draw = False
         self.gun_target = (0, 0)
         self.draw_speed = random.randint(0, 6)
         self.c_draw = 0
 
         self.gunshot_sound = pygame.mixer.Sound(os.path.join(ASSETS_DIRECTORY, SOUNDS_DIRECTORY, "gunshot.wav"))
-        self.spawned_entities = []
-        self.last_shot = int(time.time())
 
+        self.last_shot = int(time.time())
         self.gun_cooldown = random.randint(0, 2)
         self.accuracy = random.randint(-20, 20)
-        print(self.gun_cooldown)
-        print(self.accuracy)
-        print(self.draw_speed)
 
         self.direction = "left"
         self.surf = self.Animation_Idle.get_frame(position=0, direction=self.direction)
@@ -394,12 +311,6 @@ class Bandit(pygame.sprite.Sprite):
         self.rect = self.surf.get_rect(
             center=start_pos
         )
-
-        self.v = [0, 0]
-        self.acceleration = 10
-        self.gravity = 10
-        self.max_v = [50, 200]
-        self.animation_count = 0
 
     def update(self, GS, keys_pressed):
         self.idle = True
@@ -435,7 +346,7 @@ class Bandit(pygame.sprite.Sprite):
             self.update_direction("right")
         else:
             self.update_direction("left")
-        self.update_movement(GS.dt, GS.GameMap.tile_map, GS)
+        self.update_movement(GS)
 
         self.animation_count += 1
         if self.animation_count == 5:
@@ -443,49 +354,6 @@ class Bandit(pygame.sprite.Sprite):
             self.update_animation()
 
         return GS
-
-    def update_direction(self, direction="right"):
-        if self.direction != direction:
-            self.direction = direction
-            self.surf = pygame.transform.flip(self.surf, True, False)  # horizontal flip: true, vertical: false
-
-    def update_movement(self, dt, tile_map, GS):
-        if self.v[0] > self.max_v[0]:
-            self.v[0] = self.max_v[0]
-        elif self.v[0] < -self.max_v[0]:
-            self.v[0] = -self.max_v[0]
-        if self.v[1] > self.max_v[1]:
-            self.v[1] = self.max_v[1]
-        elif self.v[1] < -self.max_v[1]:
-            self.v[1] = -self.max_v[1]
-
-        if self.v[0] < 0:
-            x = -math.ceil((self.v[0] * -1) * dt)
-        else:
-            x = math.ceil(self.v[0] * dt)
-        y = math.ceil(self.v[1] * dt)
-
-        # cx, cy = [self.rect.center[0] // 16, self.rect.center[1] // 16]  # get the tile the rect is currently in
-        # print(f"The player is currently in block ({cx}, {cy})")
-        # print(f"The x velocity is {self.v[0]}\nThe applied x velocity is {x}")
-
-        if x != 0:
-            self.rect.move_ip(x, 0)
-            collisions = get_collisions(self.rect, tile_map, GS.Camera.offset)
-            for collide in collisions:
-                if x < 0:  # colliding with the right of a tile
-                    self.rect.left = collide.rect.right
-                else:  # colliding with the left of a tile
-                    self.rect.right = collide.rect.left
-        if y != 0:
-            self.rect.move_ip(0, y)
-            collisions = get_collisions(self.rect, tile_map, GS.Camera.offset)
-            for collide in collisions:
-                if y > 0:  # standing on top of a tile
-                    self.v[1] = 0
-                    self.rect.bottom = collide.rect.top
-                else:  # hitting bottom of tile
-                    self.rect.top = collide.rect.bottom
 
     def update_animation(self):  # TODO swap frame increment and display so that the first frame is displayed
         if not self.gun_draw:
@@ -512,7 +380,6 @@ class Bandit(pygame.sprite.Sprite):
         self.surf.set_colorkey((255, 255, 255))
 
     def on_hit(self):
-        print("BANDIT HIT")
         self.kill()
 
     def trigger_gunfire(self):
@@ -534,47 +401,3 @@ class Bandit(pygame.sprite.Sprite):
 
         bullet = Bullet(spos, epos, owner_id=self.id)
         self.spawned_entities.append(bullet)
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
